requirements.txt
Flask
Flask-SocketIO
eventlet
google-genai
pydantic
opencv-python-headless
Pillow  
python-dotenv